requirements_deploy.txt
# Production dependencies for Personal Health Butler Discord Bot
# Deployed to Google Cloud Run

# Core Discord
discord.py==2.6.4

# AI/ML Libraries
google-genai>=1.62.0
torch==2.6.0
torchvision==0.21.0
transformers==4.50.0
ultralytics==8.3.70
opencv-python-headless==4.11.0.86
safetensors==0.7.0
sentence-transformers==3.0.1
faiss-cpu==1.9.0
Pillow==11.1.0
rapidfuzz>=3.12.0

# Data & Database
supabase==2.28.0
SQLAlchemy==2.0.40
numpy==1.26.4
pandas==2.2.3

# HTTP & API (compatible with both google-genai and supabase)
requests==2.32.3
httpx>=0.28.1,<1.0.0
aiohttp==3.11.15

# Environment & Config
python-dotenv==1.0.1
pydantic>=2.10.0,<3.0.0
pydantic-settings>=2.8.0

# Utilities
python-dateutil==2.9.0
typing-extensions>=4.14.0
pytz==2024.2
